Update 0285-snort_decoders.xml
I personally ran into issues with my snort rules not extracting data that was necessary for better data visualizations. I modified the "snort2" decoder to now extract the actual rule description(not just the id), and to grab the srcport and dstport. 

This fix causes an issue with snort alerts that generate with no ports. Adding the "snort3" decoder above allows the decoder to still process those alerts correctly.

I wanted to submit these changes so that it might help someone else with the same issue as me sending snort alerts through syslog.